Fixed index type for mesh elements

+++ include/mdv/mesh/fwd.hpp
@@ -17,7 +17,7 @@ class TangentSpace;
 
 class UvMap;
 
-using Index        = long;
+using Index        = unsigned int;
 using IndexTriplet = std::array<Index, 3>;
 // Note: preferred std::array over Eigen::Vector to leverage unpacking features
 constexpr Index invalid_index = -1;
